yards: delete the centerline code
Every attempt at a yard centerline is gone: the bundle tracer, the
spine/skeleton types and their contraction, the forest rule, the
lateral centering, and the SPLIT pass that substituted skeletons for
in-yard weave. The overlay drops yard_spine and yard_skel, and the
console drops their layers.

What remains is the part that works: region detection, the padded hull
outline, the yard's own tracks, and entrance nodes (outline crossings
and terminal bundle ends). Nothing downstream consumed the centerlines
except the substitution pass, so SPLIT is back to its tuned behaviour
and the oracle's InYard/IsYardWay/RegionWay guards are untouched.
